refactor(accelerator): extract server::bind module from server.rs (Q2 1/n)
First module of the server.rs split: bind_with_retry + bind_with_retry_inner (the
AddrInUse wait-out-a-prior-instance retry) + their 3 tests move verbatim into
src/server/bind.rs; server.rs gets `mod bind; use bind::bind_with_retry;` and
drops the now-unused TcpListener import. start/start_https call it unchanged.

Behavior-preserving: 127 lib tests (the 3 bind tests now run as
server::bind::tests::*) + clippy -D warnings + headless all green.

Q2 continues with tls + handlers extractions; the thin router+start stays in
server.rs. RISKY-CORE-PATH → covered by the batched rc-dry-run before any cut.
